IMPL: newindex can now send references from the host Lua
Types that cannot be simply copied, will be sent as references to the underlying state.

This is good, in it simplifies a lot of things.

However, it allows the memory management story to rapidly spiral out of control. For example, taking a reference from one state, and then sending it to another is now possible.

Then closing the originating state will cause a segfault or Undefined Behaviour.

This is acceptable, because this is a _low level_ library, but it is a very sharp edge waiting to cut the programmer who never bothered to read the documentation.